Initial commit
This was tested against Windows 8.1 running PSv4 and PSv5. Testing and support will soon be added for Windows 7

Executable is in the bin folder, and an msi is available to install this tool locally, if you choose to do so. This is meant to run from a network share or USB stick for portability purposes, but I give both options.

Also included is the full Powershell Studio solution, so that you can tweak the code as you see fit if you have the software available to you.